+++ shuup/front/api/__init__.py
@@ -6,7 +6,9 @@
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
 from shuup.front.api.orders import FrontOrderViewSet
-from shuup.front.api.products import FrontProductViewSet
+from shuup.front.api.products import (
+    FrontProductViewSet, FrontShopProductViewSet
+)
 from shuup.front.api.users import FrontUserViewSet
 
 
@@ -15,6 +17,7 @@ def populate_front_api(router):
     :param router: Router
     :type router: rest_framework.routers.DefaultRouter
     """
+    router.register("shuup/front/shop_products", FrontShopProductViewSet)
     router.register("shuup/front/products", FrontProductViewSet)
     router.register("shuup/front/orders", FrontOrderViewSet)
     router.register("shuup/front/user", FrontUserViewSet)